Refactor code a bit.
